+++ ykrt/src/compile/jitc_yk/codegen/x64/mod.rs
@@ -26,7 +26,6 @@ use super::{
         jit_ir::{self, BinOp, FloatTy, Inst, InstIdx, Module, Operand, TraceKind, Ty},
         CompilationError,
     },
-    reg_alloc::{self, VarLocation},
     CodeGen,
 };
 #[cfg(any(debug_assertions, test))]
@@ -49,7 +48,7 @@ use dynasmrt::{
     dynasm,
     x64::{Rq, Rx},
     AssemblyOffset, DynamicLabel, DynasmApi, DynasmError, DynasmLabelApi, ExecutableBuffer,
-    Register,
+    Register as dynasmrtRegister,
 };
 use indexmap::IndexMap;
 use parking_lot::Mutex;
@@ -139,12 +138,143 @@ static RBP_DWARF_NUM: u16 = 6;
 /// The x64 SysV ABI requires a 16-byte aligned stack prior to any call.
 const SYSV_CALL_STACK_ALIGN: usize = 16;
 
+/// To stop us having to say `VarLocation<Register>` everywhere, we use this type alias so that
+/// within this `x64` module and its descendants we can just say `VarLocation`.
+pub(crate) type VarLocation = super::reg_alloc::VarLocation<Register>;
+
 /// A function that we can put a debugger breakpoint on.
 /// FIXME: gross hack.
 #[cfg(debug_assertions)]
 #[no_mangle]
 #[inline(never)]
 pub extern "C" fn __yk_break() {}
+
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub(crate) enum Register {
+    GP(Rq), // general purpose
+    FP(Rx), // floating point
+}
+
+impl VarLocation {
+    pub(crate) fn from_yksmp_location(m: &Module, iidx: InstIdx, x: &yksmp::Location) -> Self {
+        match x {
+            yksmp::Location::Register(0, ..) => VarLocation::Register(Register::GP(Rq::RAX)),
+            yksmp::Location::Register(1, ..) => {
+                // Since the control point passes the stackmap ID via RDX this case only happens in
+                // side-traces.
+                VarLocation::Register(Register::GP(Rq::RDX))
+            }
+            yksmp::Location::Register(2, ..) => VarLocation::Register(Register::GP(Rq::RCX)),
+            yksmp::Location::Register(3, ..) => VarLocation::Register(Register::GP(Rq::RBX)),
+            yksmp::Location::Register(4, ..) => VarLocation::Register(Register::GP(Rq::RSI)),
+            yksmp::Location::Register(5, ..) => VarLocation::Register(Register::GP(Rq::RDI)),
+            yksmp::Location::Register(8, ..) => VarLocation::Register(Register::GP(Rq::R8)),
+            yksmp::Location::Register(9, ..) => VarLocation::Register(Register::GP(Rq::R9)),
+            yksmp::Location::Register(10, ..) => VarLocation::Register(Register::GP(Rq::R10)),
+            yksmp::Location::Register(11, ..) => VarLocation::Register(Register::GP(Rq::R11)),
+            yksmp::Location::Register(12, ..) => VarLocation::Register(Register::GP(Rq::R12)),
+            yksmp::Location::Register(13, ..) => VarLocation::Register(Register::GP(Rq::R13)),
+            yksmp::Location::Register(14, ..) => VarLocation::Register(Register::GP(Rq::R14)),
+            yksmp::Location::Register(15, ..) => VarLocation::Register(Register::GP(Rq::R15)),
+            yksmp::Location::Register(x, ..) if *x >= 17 && *x <= 32 => VarLocation::Register(
+                Register::FP(super::x64::lsregalloc::FP_REGS[usize::from(x - 17)]),
+            ),
+            yksmp::Location::Direct(6, off, size) => VarLocation::Direct {
+                frame_off: *off,
+                size: usize::from(*size),
+            },
+            // Since the trace shares the same stack frame as the main interpreter loop, we can
+            // translate indirect locations into normal stack locations. Note that while stackmaps
+            // use negative offsets, we use positive offsets for stack locations.
+            yksmp::Location::Indirect(6, off, size) => VarLocation::Stack {
+                frame_off: u32::try_from(*off * -1).unwrap(),
+                size: usize::from(*size),
+            },
+            yksmp::Location::Constant(v) => {
+                // FIXME: This isn't fine-grained enough, as there may be constants of any
+                // bit-size.
+                let byte_size = m.inst(iidx).def_byte_size(m);
+                debug_assert!(byte_size <= 8);
+                VarLocation::ConstInt {
+                    bits: u32::try_from(byte_size).unwrap() * 8,
+                    v: u64::from(*v),
+                }
+            }
+            e => {
+                todo!("{:?}", e);
+            }
+        }
+    }
+}
+
+impl From<&VarLocation> for yksmp::Location {
+    fn from(val: &VarLocation) -> Self {
+        match val {
+            VarLocation::Stack { frame_off, size } => {
+                // A stack location translates is an offset in relation to RBP which has the DWARF
+                // number 6.
+                yksmp::Location::Indirect(
+                    6,
+                    -i32::try_from(*frame_off).unwrap(),
+                    u16::try_from(*size).unwrap(),
+                )
+            }
+            VarLocation::Direct { frame_off, size } => {
+                yksmp::Location::Direct(6, *frame_off, u16::try_from(*size).unwrap())
+            }
+            VarLocation::Register(reg) => {
+                let dwarf = match reg {
+                    Register::GP(reg) => match reg {
+                        Rq::RAX => 0,
+                        Rq::RDX => 1,
+                        Rq::RCX => 2,
+                        Rq::RBX => 3,
+                        Rq::RSI => 4,
+                        Rq::RDI => 5,
+                        Rq::R8 => 8,
+                        Rq::R9 => 9,
+                        Rq::R10 => 10,
+                        Rq::R11 => 11,
+                        Rq::R12 => 12,
+                        Rq::R13 => 13,
+                        Rq::R14 => 14,
+                        Rq::R15 => 15,
+                        e => todo!("{:?}", e),
+                    },
+                    Register::FP(reg) => match reg {
+                        Rx::XMM0 => 17,
+                        Rx::XMM1 => 18,
+                        Rx::XMM2 => 19,
+                        Rx::XMM3 => 20,
+                        Rx::XMM4 => 21,
+                        Rx::XMM5 => 22,
+                        Rx::XMM6 => 23,
+                        Rx::XMM7 => 24,
+                        Rx::XMM8 => 25,
+                        Rx::XMM9 => 26,
+                        Rx::XMM10 => 27,
+                        Rx::XMM11 => 28,
+                        Rx::XMM12 => 29,
+                        Rx::XMM13 => 30,
+                        Rx::XMM14 => 31,
+                        Rx::XMM15 => 32,
+                    },
+                };
+                // We currently only use 8 byte registers, so the size is constant. Since these are
+                // JIT values there are no extra locations we need to worry about.
+                yksmp::Location::Register(dwarf, 8, Vec::new())
+            }
+            VarLocation::ConstInt { bits, v } => {
+                if *bits <= 32 {
+                    yksmp::Location::Constant(u32::try_from(*v).unwrap())
+                } else {
+                    todo!(">32 bit constant")
+                }
+            }
+            e => todo!("{:?}", e),
+        }
+    }
+}
 
 /// A simple front end for the X64 code generator.
 pub(crate) struct X64CodeGen;
@@ -1109,10 +1239,10 @@ impl<'a> Assemble<'a> {
         let m = VarLocation::from_yksmp_location(self.m, iidx, self.m.param(inst.paramidx()));
         debug_assert!(self.m.inst(iidx).def_byte_size(self.m) <= REG64_BYTESIZE);
         match m {
-            VarLocation::Register(reg_alloc::Register::GP(reg)) => {
+            VarLocation::Register(Register::GP(reg)) => {
                 self.ra.force_assign_inst_gp_reg(&mut self.asm, iidx, reg);
             }
-            VarLocation::Register(reg_alloc::Register::FP(reg)) => {
+            VarLocation::Register(Register::FP(reg)) => {
                 self.ra.force_assign_inst_fp_reg(iidx, reg);
             }
             VarLocation::Direct { frame_off, size: _ } => {
@@ -1851,7 +1981,7 @@ impl<'a> Assemble<'a> {
                     size: size_dst,
                 } => {
                     match src {
-                        VarLocation::Register(reg_alloc::Register::GP(reg)) => match size_dst {
+                        VarLocation::Register(Register::GP(reg)) => match size_dst {
                             8 => dynasm!(self.asm;
                                 mov QWORD [rbp - i32::try_from(off_dst).unwrap()], Rq(reg.code())
                             ),
@@ -1896,10 +2026,10 @@ impl<'a> Assemble<'a> {
                     // match. But since the value can't change we can safely ignore this.
                 }
                 VarLocation::Register(reg) => match reg {
-                    reg_alloc::Register::GP(r) => {
+                    Register::GP(r) => {
                         gp_regs[usize::from(r.code())] = RegConstraint::InputIntoReg(op.clone(), r);
                     }
-                    reg_alloc::Register::FP(r) => {
+                    Register::FP(r) => {
                         fp_regs[usize::from(r.code())] = RegConstraint::InputIntoReg(op.clone(), r);
                     }
                 },
@@ -2020,10 +2150,10 @@ impl<'a> Assemble<'a> {
                     // Write the varlocations from the head jump to the body start.
                     // FIXME: This is copied verbatim from `cg_param` and can be reused.
                     match varloc {
-                        VarLocation::Register(reg_alloc::Register::GP(reg)) => {
+                        VarLocation::Register(Register::GP(reg)) => {
                             self.ra.force_assign_inst_gp_reg(&mut self.asm, iidx, reg);
                         }
-                        VarLocation::Register(reg_alloc::Register::FP(reg)) => {
+                        VarLocation::Register(Register::FP(reg)) => {
                             self.ra.force_assign_inst_fp_reg(iidx, reg);
                         }
                         VarLocation::Direct { frame_off, size: _ } => {
